chore: group common functionality
